Relocate numba specific tests in test_stats (and skip if not installed)
